requirements.txt
PyPDF2
pandas
yfinance
python-exchangeratesapi